cli/command/stack: internalize RunDeploy, RunRemove
These were deprecated in ad6ab189a642625d54053ac6e7080e8940288d2d and
were only used internally. Move them back inside the stack package.
