+++ cli/command/stack/deploy_composefile_test.go
@@ -1,4 +1,4 @@
-package swarm
+package stack
 
 import (
 	"context"
